Added variables to a session

+++ src/controllers/products-controller.ts
@@ -1,13 +1,15 @@
 import { Request, Response } from 'express';
-import { Product, products } from "../data/products";
+import { Types } from 'mongoose';
 import Products from '../db/schemas/product';
 
 const getProducts = async (req: Request, res: Response): Promise<void> => {
     const itemsPerPage: number = 3;
     const page: number = parseInt(req.query.page as string);
     const start: number = (page - 1) * itemsPerPage;
-    const total: number = await Products.count();
-    const products = await Products.find().skip(start).limit(itemsPerPage);
+    const total: number = await Products.count({user: req.session.userId});
+    const products = await Products.find({
+        user: req.session.userId
+    }).skip(start).limit(itemsPerPage);
     res.send({
         page: page,
         per_page: itemsPerPage,
@@ -19,7 +21,10 @@ const getProducts = async (req: Request, res: Response): Promise<void> => {
 
 const getProductById = async (req:Request, res: Response): Promise<void> => {
     const { productId } = req.params;
-    const product = await Products.findById(productId).populate({
+    const product = await Products.findOne({
+        _id: productId,
+        user: req.session.userId
+    }).populate({
         path: 'user',
         select: {
             password: 0,
@@ -34,15 +39,16 @@ const getProductById = async (req:Request, res: Response): Promise<void> => {
 };
 
 const createProduct = async (req:Request, res: Response): Promise<void> => {
-    const { name, year, price, description, user } = req.body;
-    const product = await Products.create({ name, year, price, description, user });
+    const { name, year, price, description } = req.body;
+    const { email, userId } = req.session;
+    const product = await Products.create({ name, year, price, description, user: userId });
     res.send(product);
 };
 
 const updateProduct = async (req:Request, res: Response): Promise<void> => {
     const id: string = req.params.productId;
-    const { name, year, description, price, user } = req.body;
-    const product = await Products.findByIdAndUpdate(id, { name, year, description, price, user });
+    const { name, year, description, price } = req.body;
+    const product = await Products.findOneAndUpdate({ _id: id, user: req.session.userId }, { name, year, description, price, user: req.session.userId });
     if(product) {
         res.send({data: product});
         return;
@@ -53,13 +59,12 @@ const updateProduct = async (req:Request, res: Response): Promise<void> => {
 const partialUpdateProduct = async (req:Request, res: Response): Promise<void> => {
     const id: string = req.params.productId;
     const { name, year, description, price, user } = req.body;
-    const product = await Products.findById(id);
+    const product = await Products.findOneAndUpdate({ _id: id, user: req.session.userId });
     if (product) {
         product.name = name || product.name;
         product.year = year || product.year;
         product.price = price || product.price;
         product.description = description || product.description;
-        product.user = user || product.user;
         await product.save();
         res.send({ data: product });
         return;
@@ -69,16 +74,15 @@ const partialUpdateProduct = async (req:Request, res: Response): Promise<void> =
 
 const updateProductAndNotify = async (req: Request, res: Response): Promise<void> => {
     const id: string = req.params.productId;
-    const { name, year, description, price, user } = req.body;
-    const product = await Products.findById(id);
+    const { name, year, description, price } = req.body;
+    const product = await Products.findOneAndUpdate({ _id: id, user: req.session.userId });
     if (product) {
         product.name = name || product.name;
         product.year = year || product.year;
         product.price = price || product.price;
         product.description = description || product.description;
-        product.user = user || product.user;
         await product.save();
-        res.send({ data: product, message: `Email sent to ${user.email}` });
+        res.send({ data: product, message: `Email sent to ${req.session.email}` });
         return;
     }
     res.status(404).send({});
@@ -87,7 +91,7 @@ const updateProductAndNotify = async (req: Request, res: Response): Promise<void
 const deleteProductById = async (req:Request, res: Response): Promise<void> => {
     try {
         const id: string = req.params.productId;
-        const product = await Products.deleteOne({_id: id});
+        const product = await Products.deleteOne({ _id: id, user: req.session.userId });
         if (product) {
             res.send({});
             return;
